requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
dnspython==2.6.1
typer==0.12.3
rich==13.7.1
sqlite-utils==3.36
pydantic==2.7.1
rapidfuzz==3.9.3
phonenumbers==8.13.37
python-dotenv==1.0.1
tenacity==8.3.0
tldextract==5.1.2
urllib3==2.2.1
